--- src/main/kotlin/taehyeon/brothers/matchreal/presentation/controller/daily/DailyController.kt
@@ -1,6 +1,0 @@
-package taehyeon.brothers.matchreal.presentation.controller.daily
-
-import org.springframework.web.bind.annotation.RestController
-
-@RestController
-class DailyController
