--- src/main/kotlin/taehyeon/brothers/matchreal/presentation/controller/user/UserController.kt
@@ -1,7 +1,0 @@
-package taehyeon.brothers.matchreal.presentation.controller.user
-
-import org.springframework.web.bind.annotation.RestController
-
-@RestController
-class MemberController {
-}
